net: xilinx: fix compile warning due to typecheck
This warning is introduced by the SDK commit 1b5d1c43433b3 ("net: xilinx:
axiethernet: Add check for transmit data FIFO vacancy").

Change the type of flags from "u32" to "unsigned long" to avoid the
compile warning as below:

include/linux/spinlock.h:251:3: note: in expansion of macro ‘typecheck’
  251 |   typecheck(unsigned long, flags); \
      |   ^~~~~~~~~
drivers/net/ethernet/xilinx/xilinx_axienet_main.c:1124:3: note: in expansion of macro ‘spin_lock_irqsave’
 1124 |   spin_lock_irqsave(&lp->ptp_tx_lock, flags);
      |   ^~~~~~~~~~~~~~~~~

+++ drivers/net/ethernet/xilinx/xilinx_axienet_main.c
@@ -1067,7 +1067,7 @@ static int axienet_create_tsheader(u8 *buf, u8 msg_type,
 #endif
 	u64 val;
 	u32 tmp[MRMAC_TS_HEADER_WORDS];
-	u32 flags;
+	unsigned long flags;
 	int i;
 
 #ifdef CONFIG_AXIENET_HAS_MCDMA
